chore: move new TrackerRelationshipCriteria to tracker package
